Foghorn.Data.Comp.Anncomp

Description

This module defines the variable set for Compustat’s ExecuComp Annual Compensation (ANNCOMP) database, represented as a GADT (AnncompV). Each constructor corresponds to a named variable and is indexed by an EconType (e.g., Float_, Int_, Text_, Date_).

Documentation

data AnncompV (i :: EconIndices) (a :: EconType) where

Generalized algebraic data type (GADT) of ExecuComp Annual Compensation variables.

Constructors

Address :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Street Address

Age :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Executive's Age

Allothpd :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
All Other Paid

Allothtot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
All Other Total

Becameceo :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Date_
Date Became CEO

Bonus :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Bonus ($)

Ceoann :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Annual CEO Flag

Cfoann :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Annual CFO Flag

ChgCtrlPymt :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Estimated Payments in event of change in control

City :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
City

Comment :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Textual Note Describing Unusual Situation

Coname :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Company Name

CoPerRol :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'CoPerRol_
ID number for each executive/company combination

Cusip :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
CUSIP and Issue Number

Datadate :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Datadate_
Datadate

DeferBalanceTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Aggregate Balance in Deferred Compensation Plans at Fiscal Year End ($)

DeferContribCoTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Registrant Contributions to Deferred Compensation Plans in Last Fiscal Year ($)

DeferContribExecTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Executive Contributions to Deferred Compensation Plans in Last Fiscal Year ($)

DeferEarningsTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Aggregate Earnings in Deferred Compensation Plans in Last Fiscal Year ($)

DeferRptAsCompTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Portion of Deferred Earnings Reported As Compensation ($)

DeferWithdrTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Aggregate Withdrawals/Distributions From Deferred Compensation Plans in Last Fiscal Year ($)

EipUnearnNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Equity Incentive Plan--Number of Unearned Shares That Have Not Vested

EipUnearnVal :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Equity Incentive Plan - Value of Unearned/Unvested Shares at Fiscal Year End ($)

Exchange :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Stock Exchange Company Trades on

Execdir :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Executive served as a director during the fiscal year

Execid :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Executive ID number

Execrank :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Current Rank by Salary + Bonus

Execrankann :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Executive Rank by Salary + Bonus

ExecFname :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
First Name

ExecFullname :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Full Name

ExecLname :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Last Name

ExecMname :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Middle Name

Gender :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Gender

Gvkey :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Gvkey_
Company ID Number

Inddesc :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Industry Group Description

Interlock :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Executive Is Listed in the Compensation Committee Interlocks section of the proxy

JoinedCo :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Date_
Date Joined Company

Leftco :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Date_
Date Left Company

Leftofc :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Date_
Date Left as CEO

Ltip :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
LTIP Payouts

Naics :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
NAICS Code

Naicsdesc :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Description of NAICS Code

Nameprefix :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Name Prefix

NoneqIncent :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Non-Equity Incentive Plan Compensation ($)

OldDatafmtFlag :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
True Indicates data in this table is based on the 1992 reporting rules

OptionAwards :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Value of Option Awards - FAS 123R ($)

OptionAwardsBlkValue :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Options Granted ($ - Compustat Black Scholes value)

OptionAwardsFv :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Grant Date Fair Value of Options Granted ($ - as valued by company)

OptionAwardsNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Options Granted

OptionAwardsRptValue :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Options Granted ($ - As Reported by Company)

OptExerNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Number of Shares Acquired on Option Exercise

OptExerVal :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Value Realized on Option Exercise ($)

OptUnexExerEstVal :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Estimated Value of In-the-Money Unexercised Exercisable Options ($)

OptUnexExerNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Unexercised Exercisable Options

OptUnexUnexerEstVal :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Estimated Value Of In-the-Money Unexercised Unexercisable Options ($)

OptUnexUnexerNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Unexercised Unexercisable Options

Othann :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Other Annual

Othcomp :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
All Other Compensation ($)

Page :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Present Age

Pceo :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Current CEO

Pcfo :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Current CFO

PensionChg :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Change in Pension Value and NonQualified Deferred Compensation Earnings ($)

PensionPymtsTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Payments Made From All Pension Plans During the Year ($)

PensionValueTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Present Value of Accumulated Pension Benefits From All Pension Plans ($)

Reason :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Reason Left Company

Rejoin :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Date_
Date Rejoined Company

Releft :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Date_
Date Releft Company

Reprice :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Executive Is Listed in a Stock Option Reprice table

RetYrs :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Years Credited in Company's Retirement Plan

Rstkgrnt :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Restricted Stock Grant ($)

Rstkvyrs :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Years Until Restricted Stock Grant Begins to Vest (if <3)

Salary :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Salary ($)

SalPct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Salary Percent Change Year-to-Year (%)

ShrownExclOpts :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Shares Owned - Options Excluded

ShrownExclOptsPct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Percentage of Total Shares Owned - Options Excluded

ShrownTot :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Shares Owned - As Reported

ShrownTotPct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Percentage of Total Shares Owned - As Reported

ShrsVestNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Number of Shares Acquired on Vesting

ShrsVestVal :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Value Realized on Vesting ($)

Sic :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
SIC Code

Sicdesc :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
SIC Code Description

Spcode :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
S&P Index

Spindex :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Industry Group

State :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
State

StockAwards :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Value of Stock Awards - FAS 123R ($)

StockAwardsFv :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Grant Date Fair Value of Stock Awarded Under Plan-Based Awards ($)

StockUnvestNum :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Restricted Stock Holdings

StockUnvestVal :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Restricted Stock Holdings ($)

SubTele :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Area Code

Tdc1 :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation (Salary + Bonus + Other Annual + Restriced Stock Grants + LTIP Payouts + All Other + Value of Option Grants)

Tdc1RankAnnual :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Annual ranking by TDC1 Total Compensation

Tdc2 :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation (Salary + Bonus + Other Annual + Restriced Stock Grants + LTIP Payouts + All Other + Value of Options Exercised)

Tdc1Pct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
TDC1 Percent Change Year-to-Year (%)

Tdc2Pct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
TDC2 Percent Change Year-to-Year (%)

Tele :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Telephone Number

TermPymt :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Estimated Payments in event of involuntary termination

Ticker :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Ticker Symbol

Title :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Most Recent Title

Titleann :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Annual Title

TotalAlt1 :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation - Stock/Options Valued Using Grant Date Fair Value ($)

TotalAlt2 :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation - Stock Valued at time of vesting/Options Valued at Time of Exercise ($)

TotalAlt1Pct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation (fair value of stock/option awards) Percent Change Year-to-Year (%)

TotalAlt2Pct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation (options exercised/stock vested) Percent Change Year-to-Year (%)

TotalCurr :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Current Compensation (Salary + Bonus)

TotalCurrPct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Current Compensation Percent Change Year-to-Year (%)

TotalSec :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation - As Reported in SEC Filings ($)

TotalSecPct :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Float_
Total Compensation (SEC total) Percent Change Year-to-Year (%)

Year :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Int_
Fiscal Year

Zip :: AnncompV ('Panel 'Gvkey_ 'Datadate_) 'Text_
Zip Code